CABINET: Removed HttpManagerBl
- Single method executing HTTP requests moved to
  AbstractPublicationSystemStrategy. Each strategy implementation
  now extends this abstract class.
- Removed HttpManagerBl including all references.
- Refactored test class names, removed dummy test preventing
  "no runnable methods" error (class is now abstract).
