+++ perun-cabinet/src/main/java/cz/metacentrum/perun/cabinet/strategy/impl/MUStrategy.java
@@ -5,7 +5,8 @@ import cz.metacentrum.perun.cabinet.model.Publication;
 import cz.metacentrum.perun.cabinet.model.PublicationSystem;
 import cz.metacentrum.perun.cabinet.bl.CabinetException;
 import cz.metacentrum.perun.cabinet.bl.ErrorCodes;
-import cz.metacentrum.perun.cabinet.strategy.IFindPublicationsStrategy;
+import cz.metacentrum.perun.cabinet.strategy.AbstractPublicationSystemStrategy;
+import cz.metacentrum.perun.cabinet.strategy.PublicationSystemStrategy;
 import cz.metacentrum.perun.core.api.exceptions.InternalErrorException;
 import cz.metacentrum.perun.core.impl.Utils;
 import org.apache.commons.lang.WordUtils;
@@ -49,7 +50,7 @@ import java.util.Map;
  *
  * @author Pavel Zlamal <zlamal@cesnet.cz>
  */
-public class MUStrategy implements IFindPublicationsStrategy {
+public class MUStrategy extends AbstractPublicationSystemStrategy {
 
 	private Logger log = LoggerFactory.getLogger(MUStrategy.class);
 
@@ -228,36 +229,6 @@ public class MUStrategy implements IFindPublicationsStrategy {
 
 		return publication;
 
-	}
-
-	/**
-	 * Get xml Node and xpath expression to get value from node by this xpath.
-	 *
-	 * @param node node for getting value from
-	 * @param xpathExpression expression for xpath to looking for value in node
-	 * @param resultType type of resulting / expected object (string number node nodelist ...)
-	 * @return object extracted from node by xpath
-	 * @throws InternalErrorException
-	 */
-	private Object getValueFromXpath(Node node, String xpathExpression, QName resultType) throws InternalErrorException {
-		//Prepare xpath expression
-		XPathFactory xPathfactory = XPathFactory.newInstance();
-		XPath xpath = xPathfactory.newXPath();
-		XPathExpression expr;
-		try {
-			expr = xpath.compile(xpathExpression);
-		} catch (XPathExpressionException ex) {
-			throw new InternalErrorException("Error when compiling xpath query.", ex);
-		}
-
-		Object result;
-		try {
-			result = expr.evaluate(node, resultType);
-		} catch (XPathExpressionException ex) {
-			throw new InternalErrorException("Error when evaluate xpath query on node.", ex);
-		}
-
-		return result;
 	}
 
 	/**
